Update patch set 3
Patch Set 3:

I have implemented Marc-André's suggestion to have Binary.isExecutable() return true in case of a shared object which has a main function. This has the benefit that in the project explorer the binary is shown with the executable icon. Not sure if there's any other side-effect.

Patch-set: 3
Label: Verified=0