Update patch set 2
Patch Set 2:

> With Photon RC4a for CPP I added http://download.eclipse.org/staging/photon
 > and the repo for this patchset to available repos and this is the
 > error when I try to install the upgraded version:

Oh, we need to add the javax plugins to the CDT repo so that they end up on the simrel. Usually it doesn't matter because chances are other projects will already have the orbit plugin in their repo but that is not the case here. Also...they are needed in the stand-alone debugger!

Patch-set: 2
Label: Verified=0