Update patch set 3
Patch Set 3:

(3 comments)

Thanks for the patch!

Please add a new test case to SemanticHighlightingTest that exercises the "explicit base type but not an enum class" case among others.

You'll also need to adjust the expected highlightings for some of the existing test cases (you currently have a failing test because of that).

Patch-set: 3
Label: Code-Review=0