Update patch set 1
Patch Set 1: Code-Review-1

(6 comments)

(I have retriggered the builds as the failure seems a temporary network problem @eclipse. I happen to know they are installing new network hardware today, so perhaps that is the cause?)

This looks to be an interesting patch. My main issue is that there is nothing that exercises this code. Can you provide any test code please? the org.eclipse.cdt.ui.tests plug-in gives you an opportunity to contribute example/test implementations of extension points. I would expect that at a minimum so that we can manually test going forward. Even better would be to automate that test, perhaps with SWTBot in this case. WDYT?

Patch-set: 1
Label: Code-Review=-1