Update patch set 1
Patch Set 1:

There is a problem with this approach: FieldEditorPreferencePage only actually modifies the preferences (and thus triggers property change events for the preference changes) when 'Apply' (or 'Apply and Close') is clicked.

What we would like, is for the path to be updated as soon as the radio button selection is changed, without waiting for the user to press 'Apply'.

It looks like FieldEditor does generate property change events of a different kind whenever the value of an editor changes: it generates property change events whose getProperty() is FieldEditor.VALUE, and whose getSource() is the FieldEditor object that changed. We could listen to property change events of this kind, and update the path in response.

FieldEditorPreferencePage is already an IPropertyChangeListener registered to receive property change events of this kind, so we can simply override propertyChange() in our derived class (CPPLanguageServerPreferencePage).

The other piece of the puzzle, is getting the FieldEditor for the server path to update what is shows after we modify the server path preference. This can be accomplished using FieldEditor.load().

Patch-set: 1
Reviewer: Gerrit User 973 <973@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=0