Update patch set 5
Patch Set 5:

Now getting the list of symbols from the IBinaryObject, as the children of Binary are compile units in case debug info is present (and the preference option to show them is true).

This seems to work both for release and debug builds.

I would cache the value for future use to optimise execution time, but I'm not sure when/if it should be refreshed.

Patch-set: 5
Reviewer: Gerrit User 1372 <1372@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0