Update patch set 1
Patch Set 1:

(3 comments)

Thanks for the patch!

Using StatusLineManager.setMessage() will show the message in the main status field of the status line. This status field is shared by all plugins, which means that if multiple plugins try to use it for a persistent status message like we are doing, they will be stepping on each other's toes. I don't think we should do this.

Rather, we should create and use a separate status field for these messages. This can be done something like this:

  // SETUP:

  // This is an ID that we get to make up.
  // We usually prefix such IDs with the plugin name to 
  // ensure they won't conflict with IDs used by other
  // plugins.
  String statusFieldId = "org.eclipse.lsp4e.cpp.status";

  int width = 25;  // in characters

  IStatusField statusField = new StatusLineContributionItem(
      statusFieldId, true, width);

  statusLineManager.add(statusField);

  // USAGE:

  statusField.setText(message);

The "setup" part only needs to be done once per window.

Patch-set: 1
Label: Code-Review=0