Update patch set 4
Patch Set 4:

What I am looking for is something in between. The tree should be created synchronously, that is a fast operation. Then, async, the files should be checked for existence. So when the tree is initially created, it should have the existence flag set to "unknown", and once the async code runs it should set the flag to exists/not-exists. Then the viewer should be refreshed, explicitly passing the node that is updated to the refresh method of the viewer so that not all the tree needs to be refreshed.

Patch-set: 4