Update patch set 4
Patch Set 4:

> What I am looking for is something in between. The tree should be
 > created synchronously, that is a fast operation. Then, async, the
 > files should be checked for existence. So when the tree is
 > initially created, it should have the existence flag set to
 > "unknown", and once the async code runs it should set the flag to
 > exists/not-exists. Then the viewer should be refreshed, explicitly
 > passing the node that is updated to the refresh method of the
 > viewer so that not all the tree needs to be refreshed.

So I've added the asynchronous update of the tree leaves and changes the file exist field in the DebugTree from boolean to an enum. Now, when I run the code, and click on the button "display only existing files" I have nothing that gets displayed and I suspect it is due to the fact that the folder nodes (root included) have UNKNOWN as a file exist field whereas before it all used to be set to true. 
I might to change something in the COntentProvider so that it will display the parent folders of an existing leaf node.

Patch-set: 4