Update patch set 5
Patch Set 5:

(3 comments)

Hi Baha, I think you have code in threads backwards. It looks to me that the tree is created in the job, but the existence check is in main thread. If so, you need to invert that, or at least move existence check into job too.

One extra thing that needs to be done is cancel job if a new call to displaySourceFiles is made before the last one finishes. This can happen if the user is single stepping quickly.

Patch-set: 5