Update patch set 1
Patch Set 1:

Hi Doug, I actually created a 9.5.x patch.  I couldn't think of a way to do it without changing API which means it can't go into a 9.5.x release.  I would much prefer to make a flag in the build configuration that can be set via a public method.  This can be done for 9.6.0

Patch-set: 1