Update patch set 5
Patch Set 5:

(8 comments)

Thanks, this is looking much better!

I notice that we never remove mappings from fileInactiveRegionsMap. I thought about where we could do this, but didn't have any good ideas. (Removing the mapping corresponding to the old document in inputDocumentChanged() almost works, but if the file is open in multiple windows, the same document object is shared by both editors, and we don't want to remove the mapping if only one of them is closed.)

However, now that the value type of the map is a List<Region>, leaking some mappings is not a big deal. (Previously, when the value type was the line background listener, leaking it could cause the entire editor part to stay alive.) So I think this is fine. Perhaps we could add a TODO comment above the declaration of the map, in case we think of something later.

Patch-set: 5
Label: Verified=0