Update patch set 4
Patch Set 4: Code-Review-2

(10 comments)

(Thanks for the ping, I had started reviewing it and then it never made it to the top of my list again.)

This is pretty close to right. The tests make it much better and I can experiment with it due to the decl in the plugin.xml. Thanks.

Some minor things in here need to be addressed, and one significant one. I have a real issue with the waitForDone use. However if you can explain it to me I am persuadable.

Patch-set: 4
Label: Code-Review=-2