Update patch set 1
Patch Set 1:

I believe I discovered the reason why we are seeing the behaviour where semantic highlighting sometimes disables the lexical highlighting.

PresentationReconciler is designed so that it doesn't update the entire document's presentation after every change. Rather, in response to each change, it computes a region for which the presentation needs to be updated, and only updates that region. This region is called the "damage" region, and is passed in as the first argument to createPresentation().

Accordingly, PresentationReconciler.createPresentation() only adds those lexical highlighting ranges to the presentation which overlap the damage region.

What we are doing, is adding _all_ semantic highlighting ranges to the presentation, regardless of whether they overlap the damage region.

As a result, in cases where the damage region is smaller than the whole document (which is most cases), the resulting presentation has lexical ranges for just the damage region, but semantic ranges for the whole document.

The reconciler computes the "extent" of the presentation, which is the region encompassing all the ranges in the presentation, and updates that part of the document with the ranges. Since we have added semantic ranges for the whole document, the extent will be the whole document, and therefore the whole document's style ranges will be replaced with the style ranges in the presentation. However, those style ranges are missing the lexical ranges from the parts of the document that were outside the damage region.

As a result, parts of the document outside of the damage region lose their lexical highlighting.

The solution is for our override of createPresentation() to respect the contract of that function: only add those semantic ranges to the presentation, which overlap the damage region. Then, the "extent" of the presentation will be just the damage region, and only that part of the document's style ranges will be updated, and lexical ranges outside of that region will be preserved from previous reconciles that touched those areas.

Let me know if that makes sense.

Patch-set: 1