Update patch set 1
Patch Set 1:

A numerical example may be helpful to illustrate the issue.

Suppose a document has lexical ranges at [20,25] and [70,75], and semantic ranges at [5,10], [50,55], and [100,105].

Suppose the damage region is [50,80].

Of the lexical ranges, PresentationReconciler will only add [70,75] to the presentation, because that's the only one that overlaps the damage region.

However, we add all the semantic ranges: [5,10], [50,55], and [100,105].

The resulting presentation therefore has four ranges: [5,10], [50,55], [70,75], and [100,105].

The "extent" of this presentation is [5,105].

The reconciler then says, "Oh, the presentation has been updated in the range [5,105], let me redraw that area." It replaces all the currently drawn style ranges in the range [5,105], with the four that are in the presentation: [5,10], [50,55], [70,75], and [100,105].

The result is that the lexical range [20,25] has been lost.

The correct behaviour here would be for us to only add the semantic range [50,55] to the presentation, because it is the only one that overlaps the damage region. As a result, the two ranges added to the presentation will be just [50,55] and [70,75]. The "extent" of the resulting presentation will be [50,75], and only that area will be redrawn. The existing ranges outside of that area ([5,10], [20,25], and [100,105]), which have been drawn during some previous reconcile, are preserved.

Patch-set: 1
Label: Code-Review=0