Update patch set 5
Patch Set 5:

(7 comments)

Thanks! I tried the patch and it is working pretty well.

Our implementation does suffer from the same problem as the inactive regions: the ranges are not shifted when you type, so the coloring looks broken until the next file-save (which triggers cquery to send a new message). This is the same issue as bug 537102, and the solution will be the same for both: use Position objects installed on the document for the highlighting ranges. We can deal with both issues in bug 537102 (feel free to edit the bug title).

Patch-set: 5
Label: Verified=0