Update patch set 1
Patch Set 1: Code-Review-1

Using simply "sh" requires Windows to have a "sh" binary installed. Which is not always the case.

I've got a change that seems to be working (on Linux or Windows) except for meson installed by pip (ie: it work with meson.exe but not meson.py). I simply removed the /bin/sh completely.

Patch-set: 1
Label: Code-Review=-1