Update patch set 1
Patch Set 1: Code-Review-2

(1 comment)

I think the correct way to do what you are trying is to simply:

   mvn clean package javadoc:javadoc -Ddoclint

which is simpler than adding a profile that does it. You can also just set dolint=none all the time, not profile based.

However, even better IMHO, would be to set the javadoc linting options in the pom, something like this:


  <project>
  ...
   <reporting>
    ...
    <plugins>
      <plugin>
        <groupId>org.apache.maven.plugins</groupId>
        <artifactId>maven-javadoc-plugin</artifactId>
        <version>3.0.1</version>
        <configuration>
          <doclint>none</doclint>
        </configuration>
      </plugin>
    </plugins>
   </reporting>
   ...
  </project>

The TODO line isn't necessary as it is unlikely to ever be done. Simply say that doclint is off because javadocs are incomplete.

Please add to your commit message that you have fixed file encoding issues too.

Thanks for the fix!

Patch-set: 1
Label: Code-Review=-2