Bug 537176: add "update references" checkbox to resource rename
Add a wizard page for the rename of CNature resource files.

The "discouragedReference" is downgraded to "warning" because of the use
of some LTK internal classes. As a reference, most CDT plug-ins this
setting is ignore or warning. See Bug 538133 for changing LTK to make
it easier to do this.

Change-Id: I7b63b3fc97bddf364805c0b92c22ecaafbfbb6d1
Also-by: Baha El Kassaby <baha.elkassaby@gmail.com>
Signed-off-by: Baha El Kassaby <baha.elkassaby@gmail.com>