Update patch set 1
Patch Set 1:

Jonah, this test extends BaseParametrizedTestCase but is in a separate plugin due to dependency on the GDB JTAG debugging feature. What is the best way to integrate this for automated execution? Ideally, we would catch changes to the DSF and/or DSF-GDB which break GDB JTAG debugging.

Patch-set: 1
Label: Verified=0