Update patch set 3
Patch Set 3:

(14 comments)

Thanks for the patch!

This turned out to be simpler than I originally thought. By storing the data on the document (in the form of positions) rather than in objects associated with editors, like we were trying to earlier, we avoid all the issues around when an editor's document changes or when a document is open in multiple editors. I like it!

Patch-set: 3