Update patch set 4
Patch Set 4:

Note that the extra info text field indicated by Jonah does not have word wrap.

The text field has a height which would fit a couple of rows; now that it shows command lines maybe its a good idea to enable word wrap.

Otherwise the text field feels unnecessarily high. Its still possible to scroll in the text field to see the entire command, but this is also possible in the list above it.

Patch-set: 4