Update patch set 1
Patch Set 1:

Jeff, I am not in a great position to review this, it looks fine to me, but a slight flag of deadlock concerns come up when things that weren't previously synchronized become so.

There is another week until GA, so hopefully someone will identify any problem if there are any.

Patch-set: 1