Bug 525705 - Fix Spawner.waitFor(pid) returning prematurely
On Windows XP, waitFor(pid) would return -1 immediately instead of
waiting for the process to exit. This caused starting a debug session to
be stuck at "Initializing debugger services 62%" because the GDB it was
trying to talk to was terminated immediately after being started.

Original patch by ogurets Johnson <oguretsagressive@gmail.com>.

Change-Id: I532f63c7a5facdf867ed94945b0cd26b4177c3bd
Signed-off-by: Christian Walther <walther@indel.ch>