Update patch set 5
Patch Set 5:

> > Marc-Andre or Doug, if you are able to review this it would be
 > much
 > > appreciated.
 > >
 > > Hansruedi, is there any tests that can be added for this?
 > 
 > I've added some tests which I think should cover most cases.
 > 
 > The binaries are all handcrafted elf files containing no actual
 > code so they cannot actually be run.

Nice to have some tests on this. I understand these are handcrafted files; can you provide a script or comments in the test on how they are created? E.g. If you used objcopy to make them, can you provide command lines.
I think it is great to have the intent of this code verified with tests, so thanks for the effort.

Patch-set: 5
Label: Verified=0