Update patch set 1
Patch Set 1:

CContentTypes.getContentType uses the project nature to boost the CDT content types. Some  codan tests use c++ file names but not a C++ nature for the project. Adding the nature in codan tests is done by overriding isCpp(). If I do that, a lot of tests pass. Interestingly, a bunch of tests don't pass now that there is the c++ nature, regardless of the presence of the lsp4e-cpp plugin. I haven't looked into it deeply yet but it's possible that those are legitimate failures when the checkers run inside a c++ nature project.

Patch-set: 1
Label: Code-Review=0