Bug 539882: Applicability calculator should be called with option category
In some situations, it's convenient to create option categories on the
abstract tool and use an option applicability calculator to hide the
option category if no option is added to the category in an extending
instance of the tool.
To make the decision, the calculator needs to know what category is
to be checked.

Change-Id: Id4a269ee583f1f3dee88dee797ba89a0e92d1a42
Signed-off-by: Torbjörn Svensson <torbjorn.svensson@st.com>