Update patch set 1
Patch Set 1:

The general principle when working with templates is that when something (like a type) is dependent on a template parameter, we need to be conservative and assume that instantiating the template will produce valid code.

In the case of ReturnChecker, if the return type is dependent on a template parameter, as far as we know it _could_ be void, or it could be non-void, so we can't give a warning.

But in a case like this, where the return type is not dependent on a template parameter - here, it's just "void" - we can proceed the same as if the function were not a template.

You can test whether a type is dependent on a template parameter using CPPTemplates.isDependentType(IType).

Patch-set: 1