Update patch set 1
Patch Set 1:

Taking a closer look, I realized that ReturnChecker wasn't actually computing the return type as an IType in the general case.

I changed this in https://git.eclipse.org/r/130895 - it now computes the return type as an IType. (What it was doing before was silly because it was duplicating some type resolution logic that already existing elsewhere.)

Anyways, my suggestion in the previous comment should make more sense now :)

Patch-set: 1