Update patch set 1
Patch Set 1:

(1 comment)

> (1 comment)

Thanks for you comment. I am actually not very happy with this fix, but thought I'll start with something to improve on. Currently the build will silently fail, which is very confusing...

I agree, it should be at least an else if, since my error message should be the last resort...

I'll try to find if there is a missing check as you suggested.

Patch-set: 1