Bug 533588 - respect OS-specific override in platform.txt
HierarchicalProperties already allow for OS-specific overrides, so

1 - convert platformProperties to HierarchicalProperties
2 - change HierarchicalProperties to allow overwriting of non-empty
properties. Before it would only use the platform child if the value for
the entry itself was null. With this change, if a platform child exists
its value is used if
- it contains a property or
- has an empty property and no further children. In this case the value
was specifically overridden with null

This change also allows proper parsing of esp32 platform.txt

Change-Id: Ie13a1bee57c216f6ab37cfc0208ef7711dfc62e8

Change-Id: Ie13a1bee57c216f6ab37cfc0208ef7711dfc62e8
Signed-off-by: Lutz Hamann <lhamann@planettime.de>