Update patch set 1
Patch Set 1:

I am ok with this, in that it causes no harm to CDT. 

However does this really work in your case? You have only changed one of the 30+ places the getService(IGDBBackend.class) call is made. 

Does your new GDB session have a subclass of GdbDebugServicesFactory with this in it:


	@Override
	protected IDsfDebugServicesFactory newServiceFactory(ILaunchConfiguration config, String version) {
		return new MyCustomDebugServicesFactory();
	}

And then MyCustomDebugServicesFactory is a subclass of GdbDebugServicesFactory with this in it:


	@Override
	protected IMIBackend createBackendGDBService(DsfSession session, ILaunchConfiguration lc) {
		return new MyCustomBackend(session, lc);
	}

And you connect that all in with a custom GdbLaunchDelegate launch that has this in it:


	@Override
	protected IDsfDebugServicesFactory newServiceFactory(ILaunchConfiguration config, String version) {
		return new MyCustomDebugServicesFactory();
	}

Doing that gives you two advantages, 1. It means all users of getService(IGDBBackend.class) will get the right case. 2. you will have the scaffolding in place already for when the next service required is needed for this gdb.


PS It would be great to use other language than "slave": https://medium.com/@mikebroberts/let-s-stop-saying-master-slave-10f1d1bf34df

Patch-set: 1