Update patch set 1
Patch Set 1:

> I am ok with this, in that it causes no harm to CDT.
 > 
 > However does this really work in your case? You have only changed
 > one of the 30+ places the getService(IGDBBackend.class) call is
 > made.

What we would like to do is have exactly the same behavior as a "vanilla" CDT with the option of having a secondary GDB session where we can monitor the system while it's running.
Before you shout non-stop mode, this is unfortunately not an option in this case since it breaks other parts for us. The only way we can complete our function is to have 2 separate GDB client sessions to the same GDB server.

What we need is to have one GDB client controlling execution (the current implementation in CDT fulfills this) and another GDB client for monitoring the system, both in the same launch. To solve this, we have created a parallel IGDBBackend instance for the secondary channel and a parallel ICommandControl with different service names. The views that are for monitoring will lookup the non-standard service name for the ICommandControl and uses it to communicate with the GDB server.

 > Does your new GDB session have a subclass of GdbDebugServicesFactory
 > with this in it:
 > 
 > 
 > @Override
 > protected IDsfDebugServicesFactory newServiceFactory(ILaunchConfiguration
 > config, String version) {
 > return new MyCustomDebugServicesFactory();
 > }
 > 
 > And then MyCustomDebugServicesFactory is a subclass of
 > GdbDebugServicesFactory with this in it:
 > 
 > 
 > @Override
 > protected IMIBackend createBackendGDBService(DsfSession session,
 > ILaunchConfiguration lc) {
 > return new MyCustomBackend(session, lc);
 > }
 > 
 > And you connect that all in with a custom GdbLaunchDelegate launch
 > that has this in it:
 > 
 > 
 > @Override
 > protected IDsfDebugServicesFactory newServiceFactory(ILaunchConfiguration
 > config, String version) {
 > return new MyCustomDebugServicesFactory();
 > }
 > 
 > Doing that gives you two advantages, 1. It means all users of
 > getService(IGDBBackend.class) will get the right case. 2. you will
 > have the scaffolding in place already for when the next service
 > required is needed for this gdb.

This approach would replace all references to the CDT provided IGDBBackend instance. This not what we want, we want to have two at the same time.

 > PS It would be great to use other language than "slave":
 > https://medium.com/@mikebroberts/let-s-stop-saying-master-slave-10f1d1bf34df

OK, I agree, "slave" might be a bad word here.

Patch-set: 1
Reviewer: Gerrit User 178588 <178588@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0