Update patch set 1
Patch Set 1: -Code-Review

OK. Sounds like a nice little challenge you have and I don't want CDT's implementation to get in your way. So have at it :-) 

This will be in CDT 9.6.0, which has RC2 Monday and releases on Dec 19th. I am pleased to see that ST is working to keep latest CDT relevant to you.

I am just going to reword the "slave" in the commit message and then submit it.

Thanks for the patch.

Patch-set: 1
Label: Code-Review=0