Update patch set 2
Patch Set 2:

(1 comment)

We cannot be sure that the devices targeted by extenders can support extended-remote but we can provide a useful error message at launch time if an attempt to launch with extended-remote fails due to lack of support in the GDB server or a special requirement of the device which is implemented via doRemote(). We are not breaking any existing launch configurations, only providing an opportunity for some users to create a non-functional launch configuration.

I think this would be preferable to creating new versions (eg OpenOCDSocket2, GenericDevice2, ...) of all existing IGDBJtagConnection implementations in the CDT repo.

Patch-set: 2
Reviewer: Gerrit User 846 <846@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0