Update patch set 1
Patch Set 1:

Not as critical but I really don't like the acronym DAP. It's not widely used from what I've seen so I'm not sure people will know what it is. If it was me ;), I'd call it org.eclipse.cdt.debug.adapter. But that's not as big a deal as making sure debug is in the namespace.

Patch-set: 1