Update patch set 2
Patch Set 2: Code-Review-2

> If it's ok I will remove the stack management then, no problem.
 > 
 > About "crossing problem": can you provide an example? Because I
 > didn't understand exactly what you mean. CDT formatter always skips
 > the inactive sections so this code is backward compatible, but I
 > don't know if you are talking about that or not.

I was referring to inactive sections partially overlapping noformat sections. In the latest patch you are no longer checking "statement.isActive()" so this version is ok (Actually I am not sure PS1 had a problem here, but I suppose with the isActive check gone it is not a concern.)


There is a minor bug, if the file does not end in a formatter on, the whole file is formatted despite starting with a formatter off tag. This is different than JDT. 

A test is still needed for this, 1 to demonstrate it works, and 2 to make sure no one in the future unintentionally breaks it.

Finally, I don't know how I overlooked this yesterday, but there is no UI for your changes. Can you please add the corresponding preference entries?

Patch-set: 2
Label: Code-Review=-2