Update patch set 6
Patch Set 6:

Actually I thought it was good because we can format the comment itself
and *then* avoid the format. But I checked now the JDT behavior and it seems you are right, the @formatter:off row itself is already not formatter, so JDT includes even this section in the no-format region. If we want even this feature I need to look into it, but tomorrow or probably next weekend.

Patch-set: 6
Label: Verified=0