Update patch set 9
Patch Set 9:

I tried to add the test you did, but the build failed due to trailing whitespace (how did you add a tab/space without breking rules?). However now it's correct. If you want to add another test is ok. The first line is now formatted (due to trailing whitespace) only if the option is enabled, but this is compliant with JDT, even in that case using the formatter off tag but with save action trailing whitespace, the comment is "touched" automatically. So I think we are ready to merge :)

Patch-set: 9