Bug 544721 Retain rebuildState state when cloning build configs
When cloning build config, all its childrens' rebuildState and dirty
state are maintained except for the states of AdditionalInput and
InputOrder.

With this change, we also retain their rebuildState and dirty state,
if it is a clone (that is, copyId is true in the InputType).

Change-Id: I63b2d38d0065e50357887bfccb340d458d2f4441
Signed-off-by: Philip Langer <planger@eclipsesource.com>
(cherry picked from commit 90c82078ac31ecc426c5d5223b5c865dd620f2dd)