Update patch set 1
Patch Set 1:

How do you deal with adding API in internal classes on the maintenance branch (as in this change)? They seem to break the API baseline test, but I can't increase the minor version on a maintenance branch, can I?
Or do you forbid adding additional API in internal classes in CDT in general?
Thanks a lot for any input in advance!

Patch-set: 1