Update patch set 1
Patch Set 1:

Hi Philip,

There is no API for the internal classes, so totally find for this change in that regard. What failure are you seeing related to API failure. Hopefully there isn't an x-internal missing on this package export in the MANIFEST. I see Doug has already approved it for master, so cherry-picking it for older branch should be fine too.

However CDT 9.4 is currently completely unmaintained with no plans to do any further releases on it. I have kicked off the CI job to confirm that it isn't building anymore due to missing/moved/outofdate p2 locations. I am happy to review and merge changes on 9.4 if you want, but someone (you?) will have to bring the branch up to date first. 

The "code-cleanliness" failed check can be ignored, that only applies to 9.6 and master branch.

Patch-set: 1