Update patch set 1
Patch Set 1:

> Thanks a lot for your quick reply, Jonah!
 > Are there still maintenance builds for 9.4 producing build
 > artifacts that we can consume?
 > If yes, I can update the CDT 9.4 branch. If no, I'm wondering
 > whether it's worth doing it as no one (including us) will really be
 > able to consume 9.4 including this fix.
 > Thanks again!

Hi Philip, 

The CI Job is still enabled[1], but is now failing, so no new builds will be published until it is passing. There will not be any more releases of CDT 9.4 going forward, but the nightly build will keep going about its business when builds pass. The CI job does publish to download.eclipse.org[2].

I suspect it won't be hard to get 9.4 building again, it needs a review of the target file. I can find some time to do that if it is useful.

[1] https://ci.eclipse.org/cdt/view/Main%20(old%20releases)/job/cdt-9.4/
[2] https://download.eclipse.org/tools/cdt/builds/9.4/nightly/

Does that help?
Jonah

Patch-set: 1