Update patch set 5
Patch Set 5:

> Can anyone help me? I don't understand if I need to change every
 > single MANIFEST.MF

this is unrelated to your changed. This complete failure in versions is related to the branch. This is the second time I have branched CDT since we changed to enforcing version numbers properly and something has gone wrong. 

I am working on your sequence of patches on the 9.7 branch. Once resolved there I will cherry-pick (and resolve versions) back in 9.8.

Patch-set: 5
Label: Code-Review=0