Update patch set 1
Patch Set 1:

Given that a solution that will fix the original testcase is involved, and 9.7 goes to RC2 on Monday (which is now the last build given our shortened release schedule) -- how would you feel about backing out bug 543913 from the 9.7 branch? We can leave it in the tree on the master branch, and work on fixing the noexcept issue properly for 9.8.

My rationale for proposing this is that, from the point of view of the user community, I think having some false negatives which have been around since forever, is less painful than having to deal with new false positives on relatively straightforward code using the standard library.

I'm curious to hear your thoughts as well.

Patch-set: 1