Update patch set 1
Patch Set 1:

Thanks for the detailed answer. I agree, it makes sense to move 543913 to the next release.

I'd like to try to implement the full functionality as it is probably a good learning experience, but I am not sure when I have time. I guess starting with (1) should be doable... Probably we should implement the c++17 wording?

Patch-set: 1