Bug 545877 - Don't take into account init var values in CFG
To perform dead code analysis or return checker analysis we can't just
take into account the initial value of a variable because the value.
Without a proper data flow analysis it's better to avoid to perform any
hypothesis on the value missing real bugs, so we check the initial
value of for const variables.


Change-Id: Ia086bce1e54524a91042637227581c48c9ae96c4
Signed-off-by: Marco Stornelli <marco.stornelli@gmail.com>