Bug 545877 - Don't take into account init var values if var is not const
Using initial value, if the variable is not const, to work out the
current value at compile-time can't have unexpected results because
we are not tracking what's going on with a proper data flow analysis.
Instead of use an heuristic we simply evaluate the value if the variable
is const. If it's not const the value is unknown.

Change-Id: Ia086bce1e54524a91042637227581c48c9ae96c4
Signed-off-by: Marco Stornelli <marco.stornelli@gmail.com>