Update patch set 3
Patch Set 3:

(13 comments)

My main high-level feedback on the strategy, is that we'll need to move away from using AST nodes and towards storing the noexcept-expression as an evaluation on the ICPPFunctionType, as mentioned before.

I think we should do this sooner rather than later, because it will change many of the EvalXXX.isNoexcept() implementations.

(A good rule of thumb is that the EvalXXX classes should never mention AST nodes, because they can be stored and read back from the index. The "IASTNode pointOfDefinition" constructor argument is an exception, but note that even that is immediately turned into an IBinding and never stored.)

I haven't looked too closely at the EvalBinding and EvalFunctionCall implementations because I feel they will change a lot as a result of the above. However, I don't understand the need for the `inCalledContext` parameter. Can you elaborate on that a bit? I feel like we should be able to avoid it.

Patch-set: 3