Update patch set 5
Patch Set 5:

I fixed the code according to the comments. However a test is failing after the change to CodeParser. With the code you posted we miss the namespace in the child class because we are just cloning the declarator. I didn't find a way to fix it until now. I think we should loop over parameters, check if a declarator is inside a namespace and add to the cloned declarator the new declSpecifier with QualifiedName. If you have an idea is welcome.

Patch-set: 5
Label: Verified=0